Alpine Docker, and single images
