+++ app/interesting_examples.json
@@ -4,10 +4,13 @@
     "Katharine_Hepburn",
     "Pika",
     "Kunsthistorisches_Museum",
+    "File:Jan_Vermeer_van_Delft_-_Young_Woman_with_a_Pearl_Necklace_-_Google_Art_Project.jpg",
+    "File:El_Tres_de_Mayo,_by_Francisco_de_Goya,_from_Prado_thin_black_margin.jpg",
     "Deconstructivism",
     "Palimpsest",
     "Victoria_and_Albert_Museum",
     "London",
     "Cat",
-    "Amedeo_Modigliani"
+    "Amedeo_Modigliani",
+    "Los_caprichos"
 ]
