+++ app/requirements.txt
@@ -1,7 +1,7 @@
-Flask
-Flask-Cache
+wheel
+logzero
+flask
 flask-cors
 iiif-prezi
 requests
 html-sanitizer
-
