+++ app/settings.py
@@ -1,0 +1,3 @@
+import os
+
+EXAMPLE_VARIABLE = os.getenv("EXAMPLE_VARIABLE")
