+++ app/templates/wiki.html
