+++ app/wsgi.py
@@ -1,0 +1,4 @@
+from main import app as application
+
+if __name__ == "__main__":
+    application.run(debug=True)
