+++ etc/nginx/default.conf
@@ -1,0 +1,10 @@
+server {
+    listen 80;
+    client_max_body_size 20M;
+    location = favicon.ico { access_log off; log_not_found off; }
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass 127.0.0.1:3000;
+    }
+}
