--- run_server.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-python -u wikipedia_to_iiif.py
-
